Reorganise test helpers: pyoxigraph subdir, TSV helpers to top-level conftest
- Move test_pyoxigraph_api.py into tests/datahandlers/pyoxigraph/ to group
  tests by external system (future home for UberGraph, ENSEMBL, etc.)
- Promote read_tsv and assert_*_file_valid from tests/datahandlers/conftest.py
  to tests/conftest.py — they were already imported by pipeline tests, so they
  belong at the top level
- Trim tests/datahandlers/conftest.py to only the pyoxigraph construction
  helpers (nn, lit, quad) that are specific to datahandler tests
