+++ tests/conftest.py
@@ -5,6 +5,59 @@ import pytest
 from src.node import NodeFactory
 from src.util import get_config
 
+
+# ---------------------------------------------------------------------------
+# TSV output assertion helpers (used by both datahandler and pipeline tests)
+# ---------------------------------------------------------------------------
+
+
+def read_tsv(path: str) -> list[list[str]]:
+    """Return non-empty lines of a TSV file split into columns."""
+    rows = []
+    with open(path) as f:
+        for line in f:
+            line = line.rstrip("\n")
+            if line:
+                rows.append(line.split("\t"))
+    return rows
+
+
+def assert_labels_file_valid(path: str) -> list[list[str]]:
+    """Assert every line is PREFIX:ID\\tLabel and return the rows."""
+    rows = read_tsv(path)
+    for cols in rows:
+        assert len(cols) == 2, f"Expected 2 columns, got {len(cols)}: {cols}"
+        assert ":" in cols[0], f"First column is not a CURIE: {cols[0]}"
+    return rows
+
+
+def assert_synonyms_file_valid(path: str) -> list[list[str]]:
+    """Assert every line is PREFIX:ID\\tlabeltype\\tLabel and return the rows."""
+    rows = read_tsv(path)
+    for cols in rows:
+        assert len(cols) == 3, f"Expected 3 columns, got {len(cols)}: {cols}"
+        assert ":" in cols[0], f"First column is not a CURIE: {cols[0]}"
+    return rows
+
+
+def assert_ids_file_valid(path: str) -> list[list[str]]:
+    """Assert every line is PREFIX:ID\\tbiolink:Category and return the rows."""
+    rows = read_tsv(path)
+    for cols in rows:
+        assert len(cols) == 2, f"Expected 2 columns, got {len(cols)}: {cols}"
+        assert ":" in cols[0], f"First column is not a CURIE: {cols[0]}"
+    return rows
+
+
+def assert_concordance_file_valid(path: str) -> list[list[str]]:
+    """Assert every line is CURIE\\trelation\\tCURIE and return the rows."""
+    rows = read_tsv(path)
+    for cols in rows:
+        assert len(cols) == 3, f"Expected 3 columns, got {len(cols)}: {cols}"
+        assert ":" in cols[0], f"First column is not a CURIE: {cols[0]}"
+        assert ":" in cols[2], f"Third column is not a CURIE: {cols[2]}"
+    return rows
+
 # Biolink Model version used throughout the test suite.  Should match config.yaml.
 BIOLINK_VERSION = get_config()["biolink_version"]
